requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cleanup==8.1.0
django-crispy-forms==2.2
pillow==10.4.0
sqlparse==0.5.0
tzdata==2024.1
